+++ event_store_on_memory.go
@@ -1,19 +1,19 @@
 package event_store_adapter_go
 
+var initialVersion = uint64(1)
+
 type EventStoreOnMemory struct {
 	events    map[string][]Event
-	versions  map[string]uint64
 	snapshots map[string]Aggregate
 }
 
 func NewEventStoreOnMemory() *EventStoreOnMemory {
-	return &EventStoreOnMemory{events: make(map[string][]Event), versions: make(map[string]uint64), snapshots: make(map[string]Aggregate)}
+	return &EventStoreOnMemory{events: make(map[string][]Event), snapshots: make(map[string]Aggregate)}
 }
 
 func (es *EventStoreOnMemory) GetLatestSnapshotById(aggregateId AggregateId) (*AggregateResult, error) {
 	snapshot := es.snapshots[aggregateId.AsString()]
 	if snapshot != nil {
-		// snapshot = snapshot.WithVersion(es.versions[aggregateId.AsString()])
 		return &AggregateResult{aggregate: snapshot}, nil
 	}
 	return &AggregateResult{}, nil
@@ -33,38 +33,32 @@ func (es *EventStoreOnMemory) PersistEvent(event Event, version uint64) error {
 	if event.IsCreated() {
 		panic("event is created")
 	}
-	if es.versions[event.GetAggregateId().AsString()] != version {
+	aggregateId := event.GetAggregateId().AsString()
+	if es.snapshots[aggregateId].GetVersion() != version {
 		return NewOptimisticLockError("Transaction write was canceled due to conditional check failure", nil)
 	}
-	var newVersion uint64
-	if event.IsCreated() {
-		newVersion = 1
-	} else {
-		newVersion = es.versions[event.GetAggregateId().AsString()] + 1
+	newVersion := initialVersion
+	if !event.IsCreated() {
+		newVersion = es.snapshots[aggregateId].GetVersion() + 1
 	}
-	es.events[event.GetAggregateId().AsString()] = append(es.events[event.GetAggregateId().AsString()], event)
-	snapshot := es.snapshots[event.GetAggregateId().AsString()]
+	es.events[aggregateId] = append(es.events[aggregateId], event)
+	snapshot := es.snapshots[aggregateId]
 	snapshot = snapshot.WithVersion(newVersion)
-	es.snapshots[event.GetAggregateId().AsString()] = snapshot
-	es.versions[event.GetAggregateId().AsString()] = newVersion
+	es.snapshots[aggregateId] = snapshot
 	return nil
 }
 
 func (es *EventStoreOnMemory) PersistEventAndSnapshot(event Event, aggregate Aggregate) error {
+	aggregateId := event.GetAggregateId().AsString()
+	newVersion := initialVersion
 	if !event.IsCreated() {
-		version := es.versions[event.GetAggregateId().AsString()]
+		version := es.snapshots[aggregateId].GetVersion()
 		if version != aggregate.GetVersion() {
 			return NewOptimisticLockError("Transaction write was canceled due to conditional check failure", nil)
 		}
+		newVersion = es.snapshots[aggregateId].GetVersion() + 1
 	}
-	var newVersion uint64
-	if event.IsCreated() {
-		newVersion = 1
-	} else {
-		newVersion = es.versions[event.GetAggregateId().AsString()] + 1
-	}
-	es.events[event.GetAggregateId().AsString()] = append(es.events[event.GetAggregateId().AsString()], event)
-	es.snapshots[event.GetAggregateId().AsString()] = aggregate.WithVersion(newVersion)
-	es.versions[event.GetAggregateId().AsString()] = newVersion
+	es.events[aggregateId] = append(es.events[aggregateId], event)
+	es.snapshots[aggregateId] = aggregate.WithVersion(newVersion)
 	return nil
 }
